requirements.txt
# --- Framework da API ---
# Para criar e rodar o servidor web
fastapi
uvicorn[standard]

# --- Machine Learning e Análise de Dados ---
# Para carregar e usar seu modelo .pkl e manipular dados
scikit-learn
pandas
numpy

# --- Análise Geográfica (Geoespacial) ---
# Para ler mapas de elevação/população e calcular áreas de dano
rasterio
geopandas
shapely

# --- Utilitários ---
# Para carregar variáveis de ambiente (chaves de API, etc.) do arquivo .env
python-dotenv
google-generativeai
requests